fix(terraform_plan): When one resource doesn't have an attribute, it should fail (#240)
Previously, when the policy wants to make sure all of the resources have
an attribute Tag, but, in the input, when one resource has the tag and the rest are not,
Tirith still mark the input as passing.

This commit makes it so that when one resource doesn't have a tag, even
though the rest has the tag. It will fail the policy.

Co-authored-by: Rafid Aslam <rafid.aslam@stackguardian.io>